Cleanup types in PC Speaker
 - use 8bit mode
 - use 16bit number of samples

+++ include/mixer.h
@@ -305,9 +305,9 @@ mixer_channel_t MIXER_AddChannel(MIXER_Handler handler, const int freq,
 mixer_channel_t MIXER_FindChannel(const char *name);
 
 /* PC Speakers functions, tightly related to the timer functions */
-void PCSPEAKER_SetCounter(int cntr, int mode);
+void PCSPEAKER_SetCounter(int cntr, uint8_t mode);
 void PCSPEAKER_SetType(bool pit_clock_gate_enabled, bool pit_output_enabled);
-void PCSPEAKER_SetPITControl(int mode);
+void PCSPEAKER_SetPITControl(uint8_t mode);
 
 // Mixer configuration and initialization
 void MIXER_AddConfigSection(const config_ptr_t &conf);
